requirements.txt
requests~=2.32.3
tk~=0.1.0
spotipy~=2.24.0
yt_dlp~=2024.10.7
mutagen~=1.47.0
python-dotenv~=1.0.1
pillow~=11.0.0
black==25.1.0